requirements.txt
appdirs==1.4.3
Django==1.11
django-crispy-forms==1.6.1
django-extensions==1.7.9
django-guardian==1.4.8
django-hamlpy==1.1
django-label-tag-attr==1.0.2
django-registration-redux==1.6
django-widget-tweaks==1.4.1
future==0.16.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
regex==2017.4.23
six==1.10.0
